allow import {convertDecorators} from 'tsickle';
